refactor(hooks): create hooks folder and move auto-animate in it
