+++ src/components/forms/ValidateCompletionsForm/ValidateCompletionsForm.tsx
@@ -1,9 +1,9 @@
 import styled from '@emotion/styled';
-import { useAutoAnimate } from '@formkit/auto-animate/react';
 import { Grid } from '@mui/material';
-import { FC, useEffect } from 'react';
+import { FC, useEffect, useRef } from 'react';
 import { Controller, useFieldArray, useForm } from 'react-hook-form';
 
+import { useAutoAnimate } from '../../../hooks/useAutoAnimate';
 import Completion from '../../../models/Completion';
 import CompletionQuantityTextField from '../completions/CompletionQuantityTextField';
 import CompletionUnitTextField from '../completions/CompletionUnitTextField';
@@ -37,7 +37,8 @@ const ValidateCompletionsForm: FC<Props> = ({ completions, formId, onSubmit }) =
     control, // control props comes from useForm
     name: 'completions'
   });
-  const [animateRef] = useAutoAnimate<HTMLFormElement>();
+  const animateRef = useRef(null);
+  useAutoAnimate(animateRef);
 
   useEffect(() => {
     const previousCompletions = getValues().completions;
