+++ src/components/settings/ExportSection.tsx
@@ -1,8 +1,8 @@
 import FileDownloadIcon from '@mui/icons-material/FileDownload';
 import { Button, Stack, Typography } from '@mui/material';
 
-import { useAppSelector } from '../../app/hooks';
 import { APP_NAME, JSON_DATA_EXTENSION } from '../../config/Constants';
+import { useAppSelector } from '../../hooks/redux';
 import { selectWholeStore } from '../../store/app/app.selector';
 import formatDate from '../../utils/formatDate';
