+++ src/pages/Statistics.tsx
@@ -2,13 +2,13 @@ import { SelectChangeEvent, Tab, Tabs, Typography } from '@mui/material';
 import { addDays, startOfDay } from 'date-fns';
 import { useState } from 'react';
 
-import { useAppSelector } from '../app/hooks';
 import MonthPanel from '../components/Statistics/MonthPanel/MonthPanel';
 import WeekPanel from '../components/Statistics/WeekPanel/WeekPanel';
 import YearPanel from '../components/Statistics/YearPanel/YearPanel';
 import TabPanel from '../components/TabPanel/TabPanel';
 import TrackerSelect from '../components/TrackerSelect/TrackerSelect';
 import { SEVEN_DAYS_AGO_DATE } from '../config/Constants';
+import { useAppSelector } from '../hooks/redux';
 import Tracker from '../models/Tracker';
 import { selectAllTrackers } from '../store/trackers/trackers.selectors';
